Separate NestedNode into standalone tree utilities (#66)
