+++ nestedtensor/csrc/jit_list_apply.cpp
@@ -1,6 +1,7 @@
 #include <jit_list_apply.h>
 #include <torch/csrc/jit/script/builtin_functions.h>
 #include <torch/extension.h>
+#include <torch/csrc/jit/pybind_utils.h>
 
 namespace torch {
 namespace nested_tensor {
