--- nestedtensor/csrc/nested_node.cpp
@@ -1,41 +1,0 @@
-#include <nested_node.h>
-#include <torch/csrc/autograd/utils/wrap_outputs.h>
-#include <torch/extension.h>
-
-#include <cstring>
-
-namespace torch {
-namespace nested_tensor {
-
-using namespace torch::jit;
-using namespace torch::autograd::utils;
-
-c10::optional<IValue> py_obj_to_ivalue(py::object py_obj) {
-  auto inferred_type = tryToInferType(py_obj);
-  if (!inferred_type.success()) {
-    return c10::nullopt;
-  }
-  auto payload = toIValue(py_obj, inferred_type.type());
-  return payload;
-}
-
-int64_t num_memory(c10::List<int64_t> size, c10::List<int64_t> stride) {
-  // 0-dim Tensors have torch.Size of .size() 0, but carry 1 memory.
-  // Empty 1-dim Tensors (torch.tensor([])) have torch.Size of .size() 1,
-  // but carry 0 memory.
-  if (size.size() == 0) {
-    return 1;
-  }
-  return size[0] * stride[0];
-}
-
-int64_t size_node_memory(SizeNode nested_size, SizeNode nested_stride) {
-  auto fn = [](c10::List<int64_t> size,
-               c10::List<int64_t> stride,
-               int64_t input) { return num_memory(size, stride) + input; };
-  return reduce<decltype(fn), int64_t, c10::List<int64_t>, c10::List<int64_t>>(
-      nested_size, nested_stride, fn, 0);
-}
-
-} // namespace nested_tensor
-} // namespace torch
